% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_caliendo_parro_2015.R
\name{caliendo_parro_2015}
\alias{caliendo_parro_2015}
\title{Compute Caliendo and Parro (2015)-type model in changes}
\usage{
caliendo_parro_2015(input, settings)
}
\arguments{
\item{input}{List of prepared initial and counterfactual conditions.}

\item{settings}{List of various settings and model dimensions}
}
\value{
List of list of new equilibrium `output`.
}
\description{
`caliendo_parro_2015()` updates the equilibrium to a counterfactual situation with new trade costs and/or other changes.
}
