% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mahlkow_wanner_2021.R
\name{carbon_policy_outputs_mw_2021}
\alias{carbon_policy_outputs_mw_2021}
\title{Carbon policy outputs of Mahlkow and Wanner (2021)}
\usage{
carbon_policy_outputs_mw_2021(
  initial_conditions,
  model_scenario,
  output,
  tariff_new0,
  export_subsidy_new0,
  supplier_expenditure_new,
  model_dimensions
)
}
\arguments{
\item{initial_conditions}{List of initial conditions as arrays.}

\item{model_scenario}{List of scenario variables as arrays, with the solved
`tariff_new`, `export_subsidy_new` and `tax_new`.}

\item{output}{List of outputs as arrays.}

\item{tariff_new0}{Array of exogenous new tariffs, dimensions: origin x destination x sector.}

\item{export_subsidy_new0}{Array of exogenous new export subsidies, dimensions: origin x destination x sector.}

\item{supplier_expenditure_new}{Matrix of new expenditure net of the tax, dimensions: country x sector.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
List `output` with the added variables.
}
\description{
Computes the outputs of [process_results()] that are specific to
[mahlkow_wanner_2021()]: tax revenue, the parts of tariff revenue and export
subsidy costs that are due to border carbon tariffs and export rebates,
carbon emissions and price levels.

Emissions are `sum_s carbon_intensity[s] * X[n, s] / (tax[n, s] * price[n, s])`,
the carbon of all goods used in country `n`: expenditure net of the tax,
divided by the price level, is the quantity used. In a club member without
other taxes, the tax revenue is the carbon tax times the emissions.
}
