% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dimnames.data.table}
\alias{dimnames.data.table}
\title{Dimension names of a data.table}
\usage{
\method{dimnames}{data.table}(d)
}
\arguments{
\item{d}{data.table}
}
\description{
Dimension names of a data.table
}
