% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dt_array_dimnames}
\alias{dt_array_dimnames}
\title{Dimension names of a data.table}
\usage{
dt_array_dimnames(d)
}
\arguments{
\item{d}{data.table}
}
\description{
Dimension names of a data.table
}
