% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_time_diff}
\alias{format_time_diff}
\title{Format Time Difference}
\usage{
format_time_diff(start_time, end_time)
}
\arguments{
\item{start_time}{POSIXct object representing the start time.}

\item{end_time}{POSIXct object representing the end time.}
}
\value{
A character string in the format "X:YY minutes" or "Z seconds".
}
\description{
Format Time Difference
}
