% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{generate_input}
\alias{generate_input}
\title{Generate inputs from initial and counterfactual conditions}
\usage{
generate_input(initial_conditions, model_scenario, settings)
}
\arguments{
\item{initial_conditions}{List of initial conditions, i.e. named data.tables}

\item{model_scenario}{List of counterfactual conditions, i.e. named data.tables}

\item{settings}{List of settings}
}
\description{
Generate inputs from initial and counterfactual conditions
}
