% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_model_dimensions}
\alias{get_model_dimensions}
\title{Get all dimensions of a model by inspecting initial conditions}
\usage{
get_model_dimensions(x)
}
\arguments{
\item{x}{List of data.tables of initial conditions}
}
\description{
Get all dimensions of a model by inspecting initial conditions
}
