% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{\%diag\%}
\alias{\%diag\%}
\title{Slightly faster implementation of diagonal elements of matrix product}
\usage{
x \%diag\% y
}
\arguments{
\item{x}{Matrix}

\item{y}{Matrix}
}
\description{
Slightly faster implementation of diagonal elements of matrix product
}
