% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{initialize_variable}
\alias{initialize_variable}
\title{Initialize variable as array}
\usage{
initialize_variable(dims, value = 1)
}
\arguments{
\item{dims}{Dimensions of variable}

\item{value}{Default value}
}
\value{
Array
}
\description{
`initialize_variable()` Initializes a variable.
}
