% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{nest_elasticity_variables}
\alias{nest_elasticity_variables}
\title{Nest elasticity-related initial conditions under a dedicated list}
\usage{
nest_elasticity_variables(x)
}
\arguments{
\item{x}{List of initial conditions or scenario inputs}
}
\description{
Nest elasticity-related initial conditions under a dedicated list
}
