% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{predict_convergence_eta}
\alias{predict_convergence_eta}
\title{Predict convergence ETA}
\usage{
predict_convergence_eta(change_list, tolerance)
}
\arguments{
\item{change_list}{A matrix with two columns:
- First column: Time (relative or absolute).
- Second column: Convergence criterion values.}

\item{tolerance}{Tolerance for which convergence is assumed to be achieved.}
}
\value{
Estimated time (in the same units as `change_list` time) to reach the `tolerance`,
or `"-"` if insufficient data.
}
\description{
Predict convergence ETA
}
