% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Process results for new equilibrium}
\usage{
process_results(results, ...)
}
\arguments{
\item{results}{List of results from `update_equilibrium()`.}

\item{...}{Currently unused.}
}
\value{
List with `initial_conditions`, `model_scenario`, `output`, and `settings`.
}
\description{
Takes the results from `update_equilibrium()` and computes additional variables.
Uses S3 dispatch on the model-specific result class.
}
