% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_expenditure.R
\name{update_expenditure}
\alias{update_expenditure}
\title{Update expenditure}
\usage{
update_expenditure(
  expenditure_new,
  consumption_share,
  input_share,
  trade_share_new,
  tariff_new,
  value_added,
  wage_change,
  trade_balance,
  model_attributes,
  verbose
)
}
\arguments{
\item{expenditure_new}{Matrix of new expenditures, dimensions: sector x country.}

\item{consumption_share}{Matrix of elasticities of substitution or consumption shares, dimensions: sector x country.}

\item{input_share}{Array of beta-multiplied input-output coefficients, dimensions: input_sector x output_sector x country.}

\item{trade_share_new}{Array of new trade flows (FOB), dimensions: origin x destination x sector.}

\item{tariff_new}{Array of new tariffs, dimensions: origin x destination x sector.}

\item{value_added}{Matrix of value added, dimension: country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{trade_balance}{Vector of aggregate trade balance, dimension: country.}

\item{model_attributes}{List of model attributes.}

\item{verbose}{verbose}
}
\value{
Matrix of new expenditures, dimensions: sector x country.
}
\description{
Updates expenditures following Caliendo & Parro (2015) equation (13).
}
