% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mahlkow_wanner_2021.R
\name{update_expenditure_mw_2021}
\alias{update_expenditure_mw_2021}
\title{Update expenditure for Mahlkow and Wanner (2021)}
\usage{
update_expenditure_mw_2021(
  expenditure_new,
  consumption_share,
  input_share,
  trade_share_new,
  tariff_new,
  export_subsidy_new,
  tax_new,
  value_added,
  wage_change,
  trade_balance_new,
  model_dimensions,
  tolerance,
  verbose,
  convergence_method,
  max_inner_iterations = 5000L
)
}
\arguments{
\item{expenditure_new}{Matrix of new expenditures, dimensions: country x sector.}

\item{consumption_share}{Matrix of consumption shares, dimensions: country x sector.}

\item{input_share}{Array of beta-multiplied input-output coefficients, dimensions: country x input x output.}

\item{trade_share_new}{Array of new trade shares, dimensions: origin x destination x sector.}

\item{tariff_new}{Array of new tariffs, dimensions: origin x destination x sector.}

\item{export_subsidy_new}{Array of export subsidies, dimensions: origin x destination x sector.}

\item{tax_new}{Matrix of new tax wedges, dimensions: country x sector.}

\item{value_added}{Vector of value added, dimension: country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{trade_balance_new}{Vector of new aggregate trade balance, dimension: country.}

\item{model_dimensions}{List of model dimensions.}

\item{tolerance}{Tolerance for inner convergence.}

\item{verbose}{Verbosity level.}

\item{convergence_method}{Method for convergence checking.}

\item{max_inner_iterations}{Maximum number of inner iterations before the loop bails out (default 5000).}
}
\value{
List with `expenditure_new` (country x sector),
  `n_iterations_expenditure_update` and `inner_converged`.
}
\description{
Solves for tax-inclusive expenditure as [update_expenditure_cp_2015()],
with suppliers receiving expenditure net of the tax wedge and the tax
revenue as part of income. [process_results()] uses it for the initial
expenditure.
}
