% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_input_cost.R
\name{update_input_cost}
\alias{update_input_cost}
\title{Update input cost matrix}
\usage{
update_input_cost(
  input_cost_change,
  wage_change,
  price_change,
  factor_share,
  intermediate_share,
  model_attributes
)
}
\arguments{
\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{factor_share}{Matrix of input factor shares, dimensions: sector x country.}

\item{intermediate_share}{Array of input-output coefficients, dimensions: input_sector x output_sector x country.}

\item{model_attributes}{List of model attributes.}
}
\value{
Matrix of changes in input costs, dimensions: sector x country.
}
\description{
Updates cost of input bundle following Caliendo & Parro (2015) equations (10).
}
