% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_chowdhry_hinz_kamin_wanner_2022.R
\name{update_input_cost_chkw_2022}
\alias{update_input_cost_chkw_2022}
\title{Update input cost matrix for Chowdhry, Hinz, Kamin & Wanner (2022)}
\usage{
update_input_cost_chkw_2022(
  input_cost_change,
  wage_change,
  price_change,
  factor_share,
  intermediate_share,
  model_dimensions
)
}
\arguments{
\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{factor_share}{Matrix of input factor shares, dimensions: sector x country.}

\item{intermediate_share}{Array of input-output coefficients, dimensions: input_sector x output_sector x country.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
Matrix of changes in input costs, dimensions: sector x country.
}
\description{
Updates cost of input bundle following Chowdhry, Hinz, Kamin & Wanner (2022) equations (10).
}
