% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_mahlkow_wanner_2021.R
\name{update_input_cost_mw_2021}
\alias{update_input_cost_mw_2021}
\title{Update input cost matrix for Mahlkow and Wanner (2021)}
\usage{
update_input_cost_mw_2021(
  input_cost_change,
  wage_change,
  productivity_change,
  price_change,
  tax_change,
  factor_share,
  intermediate_share,
  model_dimensions
)
}
\arguments{
\item{input_cost_change}{Matrix of input cost changes, dimensions: country x sector.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{productivity_change}{Matrix of change in productivity, dimensions: country x sector.}

\item{price_change}{Matrix of price index changes, dimensions: country x sector.}

\item{tax_change}{Matrix of tax wedge changes, dimensions: country x sector.}

\item{factor_share}{Matrix of input factor shares, dimensions: country x sector.}

\item{intermediate_share}{Array of input-output coefficients, dimensions: country x input x output.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
Matrix of changes in input costs, dimensions: country x sector.
}
\description{
Updates cost of input bundle as in Caliendo & Parro (2015) equation (10),
with intermediate inputs at tax-inclusive prices.
}
