% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_caliendo_parro_2015.R
\name{update_output_income_cp_2015}
\alias{update_output_income_cp_2015}
\title{Update output and income for Caliendo Parro (2015)}
\usage{
update_output_income_cp_2015(
  output_new,
  income_new,
  consumption_share,
  input_share,
  trade_share_new,
  tariff_new,
  export_subsidy_new,
  value_added,
  wage_change,
  population_change,
  trade_balance_new,
  model_dimensions,
  tolerance,
  convergence_method,
  verbose,
  max_inner_iterations = 5000L
)
}
\arguments{
\item{output_new}{Matrix of new outputs, dimensions: country x sector.}

\item{income_new}{Vector of new incomes, dimension: country.}

\item{consumption_share}{Matrix of consumption shares, dimensions: country x sector.}

\item{input_share}{Array of beta-multiplied input-output coefficients, dimensions: country x input x output.}

\item{trade_share_new}{Array of new trade shares, dimensions: origin x destination x sector.}

\item{tariff_new}{Array of new tariffs, dimensions: origin x destination x sector.}

\item{export_subsidy_new}{Array of export subsidies, dimensions: origin x destination x sector.}

\item{value_added}{Vector of value added, dimension: country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{population_change}{Vector of change in population, dimension: country.}

\item{trade_balance_new}{Vector of new aggregate trade balance, dimension: country.}

\item{model_dimensions}{List of model dimensions.}

\item{tolerance}{Tolerance for inner convergence.}

\item{convergence_method}{Method for convergence checking.}

\item{verbose}{Verbosity level.}

\item{max_inner_iterations}{Maximum number of inner iterations before the loop bails out (default 5000).}
}
\value{
Matrix of new expenditures, dimensions: country x sector.
}
\description{
Updates output and income analogous to Caliendo & Parro (2015) equation (13) for the expenditure update.
}
