% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_price.R
\name{update_price}
\alias{update_price}
\title{Update price index matrix}
\usage{
update_price(
  price_change,
  trade_share,
  trade_cost_change,
  input_cost_change,
  trade_elasticity,
  model_attributes
)
}
\arguments{
\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{trade_share}{Array of initial trade flows, dimensions: origin x destination x sector.}

\item{trade_cost_change}{Array of change in trade costs, dimensions: origin x destination x sector.}

\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{trade_elasticity}{Vector of trade elasticities, dimensions: sector.}

\item{model_attributes}{List of model attributes.}
}
\value{
Matrix of changes in price index, dimensions: sector x country.
}
\description{
Update price index following Caliendo & Parro (2015) equation (11).
}
