% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_caliendo_parro_2015.R
\name{update_price_index_cp_2015}
\alias{update_price_index_cp_2015}
\title{Update price index matrix for Caliendo and Parro (2015)}
\usage{
update_price_index_cp_2015(
  price_change,
  trade_share,
  trade_cost_change,
  input_cost_change,
  trade_elasticity,
  model_dimensions
)
}
\arguments{
\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{trade_share}{Array of initial trade flows, dimensions: origin x destination x sector.}

\item{trade_cost_change}{Array of change in trade costs, dimensions: origin x destination x sector.}

\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{trade_elasticity}{Positive Frechet trade elasticity theta, dimensions: sector.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
Matrix of changes in price index, dimensions: sector x country.
}
\description{
Update price index following Caliendo & Parro (2015) equation (11).
}
