% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_trade_share.R
\name{update_trade_share}
\alias{update_trade_share}
\title{Update trade share matrix}
\usage{
update_trade_share(
  trade_share,
  trade_cost_change,
  input_cost_change,
  price_change,
  trade_elasticity,
  model_attributes
)
}
\arguments{
\item{trade_share}{Array of initial trade flows, dimensions: origin x destination x sector.}

\item{trade_cost_change}{Array of change in trade costs, dimensions: origin x destination x sector.}

\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{trade_elasticity}{Vector of trade elasticities, dimension: sector.}

\item{model_attributes}{List of model attributes.}
}
\value{
Array of new trade shares, dimensions: origin x destination x sector.
}
\description{
Compute new trade shares following Caliendo & Parro (2015) equation (12).`
}
