% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vector_to_array}
\alias{vector_to_array}
\title{Create one-dimensional array from vector}
\usage{
vector_to_array(x, dimname = NULL)
}
\arguments{
\item{x}{vector}

\item{dimname}{dimname}
}
\description{
Create one-dimensional array from vector
}
