% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cast_variable}
\alias{cast_variable}
\title{Cast variable from data.tables to vectors, matrices and arrays}
\usage{
cast_variable(x)
}
\arguments{
\item{x}{Data.table to be reshaped into vector/matrix/array with named dimensions if it is indeed a data.frame (i.e. data.table)}
}
\value{
Vector/matrix/array
}
\description{
`cast_variable()` updates the equilibrium to a counterfactual situation with new trade costs and/or other changes.
}
