% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KITE-package.R
\docType{package}
\name{KITE-package}
\alias{KITE-package}
\title{KITE Model Suite}
\description{
A quantitative framework for international trade policy analysis.
Multi-sector Ricardian models with intra- and international input-output
linkages; supports simulation of trade policy changes via tariff and
non-tariff measure scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julianhinz/KITE}
  \item Report bugs at \url{https://github.com/julianhinz/KITE/issues}
}

}
\author{
\strong{Maintainer}: Julian Hinz \email{julian.hinz@kielinstitut.de}

Authors:
\itemize{
  \item Julian Hinz \email{julian.hinz@kielinstitut.de}
  \item Hendrik Mahlkow \email{hendrik.mahlkow@kielinstitut.de}
  \item Joschka Wanner \email{joschka.wanner@kielinstitut.de}
}

}
