% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{KITE-package}
\alias{KITE-package}
\title{KITE: Compute Counterfactual Scenarios for Trade Policy}
\description{
The KITE package provides tools to compute counterfactual scenarios for trade policy using advanced econometric methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julianhinz/KITE}
  \item Report bugs at \url{https://github.com/julianhinz/KITE/issues}
}

}
\author{
\strong{Maintainer}: Julian Hinz \email{julian.hinz@kielinstitut.de}

Authors:
\itemize{
  \item Julian Hinz \email{julian.hinz@kielinstitut.de}
  \item Hendrik Mahlkow \email{hendrik.mahlkow@kielinstitut.de}
  \item Joschka Wanner \email{joschka.wanner@kielinstitut.de}
}

}
