% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{melt_variable}
\alias{melt_variable}
\title{Melt variable from vectors, matrices and arrays, to data.tables}
\usage{
melt_variable(x)
}
\arguments{
\item{x}{vector/matrix/array to data.table}
}
\value{
Data.table
}
\description{
`melt_variable()`
}
