% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{predict_convergence_eta}
\alias{predict_convergence_eta}
\title{Predict convergence ETA}
\usage{
predict_convergence_eta(time, change, tolerance, steps = 5)
}
\arguments{
\item{time}{Vector of time, relative or absolute}

\item{change}{Vector of relative change in variable of interest}

\item{tolerance}{Tolerance for which convergence is assumed to be achieved}

\item{steps}{Steps to take into account for estimation}
}
\description{
Predict convergence ETA
}
