% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Process results for new equilibrium}
\usage{
process_results(results)
}
\arguments{
\item{results}{List of results from update_equilibrium.}
}
\value{
List of data.tables with additional variables
}
\description{
Takes the results from the updated equilibrium and computes additional variables.
}
