% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_equilibrium.R
\name{update_equilibrium}
\alias{update_equilibrium}
\title{Update equilibrium}
\usage{
update_equilibrium(
  model = NULL,
  initial_conditions = NULL,
  model_scenario = NULL,
  settings = NULL
)
}
\arguments{
\item{model}{Model specification to run, e.g. caliendo_parro_2015()}

\item{initial_conditions}{List of initial conditions.}

\item{model_scenario}{List of counterfactual conditions}

\item{settings}{List of settings}
}
\value{
List of list of data.tables with model, initial_conditions, model_scenario and output
}
\description{
`update_equilibrium()` updates the equilibrium to a counterfactual situation with new trade costs and/or other changes.
}
