% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_caliendo_parro_2015.R
\name{update_output_income_cp_2015}
\alias{update_output_income_cp_2015}
\title{Update output and income for Caliendo Parro (2015)}
\usage{
update_output_income_cp_2015(
  output_new,
  income_new,
  consumption_share,
  input_share,
  trade_share_new,
  tariff_new,
  export_subsidy_new,
  value_added,
  wage_change,
  population_change,
  trade_balance_new,
  model_dimensions,
  tolerance,
  convergence_method,
  verbose
)
}
\arguments{
\item{output_new}{Matrix of new expenditures, dimensions: sector x country.}

\item{income_new}{Matrix of new incomes, dimensions: country.}

\item{consumption_share}{Matrix of elasticities of substitution or consumption shares, dimensions: sector x country.}

\item{input_share}{Array of beta-multiplied input-output coefficients, dimensions: input_sector x output_sector x country.}

\item{trade_share_new}{Array of new trade flows (FOB), dimensions: origin x destination x sector.}

\item{tariff_new}{Array of new tariffs, dimensions: origin x destination x sector.}

\item{export_subsidy_new}{Array of export taxes, dimensions: origin x destination x sector.}

\item{value_added}{Matrix of value added, dimension: country.}

\item{wage_change}{Vector of change in wages, dimension: country.}

\item{population_change}{Vector of change in population, dimension: country.}

\item{trade_balance_new}{Vector of new aggregate trade balance, dimension: country.}

\item{model_dimensions}{List of model dimensions.}

\item{tolerance}{tolerance}

\item{convergence_method}{Method for convergence checking.}

\item{verbose}{verbose}
}
\value{
Matrix of new expenditures, dimensions: sector x country.
}
\description{
Updates output and income analogous to Caliendo & Parro (2015) equation (13) for the expenditure update.
}
