% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_commons.R
\name{update_trade_balance}
\alias{update_trade_balance}
\title{Update trade balance according to different adjustment rules}
\usage{
update_trade_balance(
  trade_balance,
  value_added,
  value_added_new,
  trade_balance_rule = "fixed"
)
}
\arguments{
\item{trade_balance}{Vector of initial trade balances, dimension: country.}

\item{value_added}{Vector of initial value added, dimension: country.}

\item{value_added_new}{Vector of new value added, dimension: country.}

\item{trade_balance_rule}{Character string specifying adjustment rule: `"fixed"`, `"fixed_country_share"`, `"fixed_global_share"`, or `"zero"`.}
}
\value{
Vector of updated trade balance values, dimension: country.
}
\description{
Updates the trade balance based on different rules:
(1) "fixed" - Unchanged trade balance,
(2) "fixed_country_share" - Constant relative to country's value added,
(3) "fixed_global_share" - Constant share in global value added,
(4) "zero" - Set all trade balances to zero.
}
