% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_chowdhry_hinz_kamin_wanner_2022.R
\name{update_trade_share_chkw_2022}
\alias{update_trade_share_chkw_2022}
\title{Update trade share matrix for Chowdhry, Hinz, Kamin & Wanner (2022)}
\usage{
update_trade_share_chkw_2022(
  trade_share,
  trade_cost_change,
  input_cost_change,
  price_change,
  trade_elasticity,
  model_dimensions
)
}
\arguments{
\item{trade_share}{Array of initial trade flows, dimensions: origin x destination x sector.}

\item{trade_cost_change}{Array of change in trade costs, dimensions: origin x destination x sector.}

\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{trade_elasticity}{Vector of trade elasticities, dimension: sector.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
Array of new trade shares, dimensions: origin x destination x sector.
}
\description{
Compute new trade shares following Chowdhry, Hinz, Kamin & Wanner (2022) equation (12).
}
