% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_caliendo_parro_2015.R
\name{update_trade_share_cp_2015}
\alias{update_trade_share_cp_2015}
\title{Update trade share matrix for Caliendo and Parro (2015)}
\usage{
update_trade_share_cp_2015(
  trade_share,
  trade_cost_change,
  input_cost_change,
  price_change,
  trade_elasticity,
  model_dimensions
)
}
\arguments{
\item{trade_share}{Array of initial trade flows, dimensions: origin x destination x sector.}

\item{trade_cost_change}{Array of change in trade costs, dimensions: origin x destination x sector.}

\item{input_cost_change}{Matrix of input cost changes, dimensions: sector x country.}

\item{price_change}{Matrix of price index changes, dimensions: sector x country.}

\item{trade_elasticity}{Vector of trade elasticities, dimension: sector.}

\item{model_dimensions}{List of model dimensions.}
}
\value{
Array of new trade shares, dimensions: origin x destination x sector.
}
\description{
Compute new trade shares following Caliendo & Parro (2015) equation (12).`
}
